Добавил зависимости loombok, jsoun,gson
