+++ src/main/java/AppLauncher.java
@@ -1,11 +1,13 @@
 
+import lombok.SneakyThrows;
 import org.telegram.telegrambots.meta.TelegramBotsApi;
-import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
-import telegramAPI.TelegramApi;
+import telegramService.TelegramApi;
 
 public class AppLauncher {
-    public static void main(String[] args) throws TelegramApiException {
+    @SneakyThrows
+    public static void main(String[] args)  {
+
 
         TelegramBotsApi botsApi = new TelegramBotsApi(DefaultBotSession.class);
         botsApi.registerBot(new TelegramApi());
